+++ lib/grape_jsonapi.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require 'jsonapi/serializer'
+require 'grape'
+require 'grape_jsonapi/endpoint_extension'
+require 'grape_jsonapi/formatter'
+require 'grape_jsonapi/parser'
+require 'grape_jsonapi/version'
+
+require 'grape_jsonapi/deprecated/formatter'
+require 'grape_jsonapi/deprecated/parser'