+++ spec/lib/grape_jsonapi/deprecated.rb/formatter_spec.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+describe Grape::Formatter::FastJsonapi do
+  subject { described_class }
+
+  it { expect(subject.deprecated?).to be true }
+end